requirements.txt
aiogram==2.23.1
flask
python-telegram-bot
python-dotenv~=1.0.1
requests
pcloud
schedule